More fixes for HVI error
There was diffusion straddling HVI and I couldn't move things around
to avoid that and also provide enough distance for HVI to other things
around.

So I just changed the tap to a LV tap and use a that LV tap to bias
substrate for both LV and MV NMOS in the area which should be fine.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>